Get the data dictionary

This function retrieves the data dictionary for the specified data directory.

Usage

spod_available_data_v2(
  data_dir = spod_get_data_dir(),
  check_local_files = FALSE,
  quiet = FALSE
)

Arguments

data_dir

The directory where the data is stored. Defaults to the value returned by spod_get_data_dir().

check_local_files

Whether to check if the local files exist. Defaults to FALSE.

quiet

A logical value indicating whether to suppress messages. Default is FALSE.

Value

A tibble with links, release dates of files in the data, dates of data coverage, local paths to files, and the download status.

target_url

character. The URL link to the data file.

pub_ts

POSIXct. The timestamp of when the file was published.

file_extension

character. The file extension of the data file (e.g., 'tar', 'gz').

data_ym

Date. The year and month of the data coverage, if available.

data_ymd

Date. The specific date of the data coverage, if available.

local_path

character. The local file path where the data is stored.

downloaded

logical. Indicator of whether the data file has been downloaded locally.

Examples

# Get the data dictionary for the default data directory
if (FALSE) {
  metadata <- spod_available_data_v2()
  names(metadata)
  head(metadata)
}