Convert all downloaded v1 origin-destination data to duckdb

Convert all downloaded v1 origin-destination data to duckdb

Usage

spod_convert_od_v1_to_duckdb(
  zones = c("districts", "dist", "distr", "distritos", "municipalities", "muni",
    "municip", "municipios"),
  data_dir = spod_get_data_dir(),
  save_dir = NULL,
  quiet = FALSE,
  duck_max_mem = 3,
  duck_max_threads = parallelly::availableCores(),
  overwrite = FALSE
)

Arguments

zones

The zones for which to download the data. Can be "districts" (or "dist", "distr", or the original Spanish "distritos") or "municipalities" (or "muni", "municip", or the original Spanish "municipios").

data_dir

The directory where the data is stored.

save_dir

The path to the directory where the duckdb files will be saved. If NULL, uses the default location in data_dir (set by the SPANISH_OD_DATA_DIR environment variable). Therefore, the default relative path is <data_dir>/clean_data/v1/tabular/duckdb/od_<zones>.duckdb.

quiet

A logical value indicating whether to suppress messages. Default is FALSE.

duck_max_mem

The maximum memory to use in GB. A conservative default is 3 GB, which should be enough for resaving the data to DuckDB form a folder of CSV.gz files while being small enough to fit in memory of most even old computers. For data analysis using the already converted data (in DuckDB or Parquet format) or with the raw CSV.gz data, it is recommended to increase it according to available resources.

duck_max_threads

The maximum number of threads to use. Defaults to the number of available cores minus 1.

overwrite

Logical. If TRUE, overwrites existing duckdb files. Defaults to FALSE.

Value

Path to saved DuckDB file.