Download the data files of specified type, zones, and dates

This function downloads the data files of the specified type, zones, dates and data version.

Usage

spod_download_data(
  type = c("od", "origin-destination", "os", "overnight_stays", "tpp",
    "trips_per_person"),
  zones = c("districts", "dist", "distr", "distritos", "municipalities", "muni",
    "municip", "municipios", "lau", "large_urban_areas", "gau", "grandes_areas_urbanas"),
  dates = NULL,
  data_dir = spod_get_data_dir(),
  quiet = FALSE,
  return_output = TRUE
)

Arguments

type

The type of data to download. Can be "origin-destination" (or ust "od"), or "trips_per_person" (or just "tpp") for v1 data. For v2 data "overnight_stays" (or just "os") is also available. More data types to be supported in the future. See respective codebooks for more information. ADD CODEBOOKS! to the package

zones

The zones for which to download the data. Can be "districts" (or "dist", "distr", or the original Spanish "distritos") or "municipalities" (or "muni", "municip", or the original Spanish "municipios") for both data versions. Additionaly, these can be "large_urban_areas" (or "lau", or the original Spanish "grandes_areas_urbanas", or "gau") for v2 data (2022 onwards).

dates

A character or Date vector of dates to process. Kindly keep in mind that v1 and v2 data follow different data collection methodologies and may not be directly comparable. Therefore, do not try to request data from both versions for the same date range. If you need to compare data from both versions, please refer to the respective codebooks and methodology documents. The v1 data covers the period from 2020-02-14 to 2021-05-09, and the v2 data covers the period from 2022-01-01 to the present until further notice. The true dates range is checked against the available data for each version on every function run.

The possible values can be any of the following:

A single date in ISO (YYYY-MM-DD) or YYYYMMDD format. character or Date object.

A vector of dates in ISO (YYYY-MM-DD) or YYYYMMDD format. character or Date object. Can be any non-consecutive sequence of dates.

A date range

eigher a character or Date object of length 2 with clearly named elements start and end in ISO (YYYY-MM-DD) or YYYYMMDD format. E.g. c(start = "2020-02-15", end = "2020-02-17");

or a character object of the form YYYY-MM-DD_YYYY-MM-DD or YYYYMMDD_YYYYMMDD. For example, 2020-02-15_2020-02-17 or 20200215_20200217.

A regular expression to match dates in the format YYYYMMDD. character object. For example, ^202002 will match all dates in February 2020.

data_dir

The directory where the data is stored. Defaults to the value returned by spod_get_data_dir() which returns the value of the environment variable SPANISH_OD_DATA_DIR or a temporary directory if the variable is not set.

quiet

A logical value indicating whether to suppress messages. Default is FALSE.

return_output

Logical. If TRUE, the function returns a character vector of the paths to the downloaded files. If FALSE, the function returns NULL.

Value

A character vector of the paths to the downloaded files. Unless return_output = FALSE, in which case the function returns NULL.

Examples

if (FALSE) { # \dontrun{
# Download the origin-destination on district level for the a date range in March 2020
spod_download_data(
  type = "od", zones = "districts",
  date_range = c("2020-03-20", "2020-03-24")
)

# Download the origin-destination on district level for select dates in 2020 and 2021
spod_download_data(
  type = "od", zones = "dist",
  dates_list = c("2020-03-20", "2020-03-24", "2021-03-20", "2021-03-24")
)

# Download the origin-destination on municipality level using regex for a date range in March 2020
# (the regex will capture the dates 2020-03-20 to 2020-03-24)
spod_download_data(
  type = "od", zones = "municip",
  date_regex = "2020032[0-4]"
)
} # }