Retrieve the origin-destination v1 data (2020-2021)

This function retrieves the v1 (2020-2021) origin-destination data from the specified data directory.

Usage

spod_get_od(
  zones = c("districts", "dist", "distr", "municipalities", "muni", "municip"),
  dates = NULL,
  data_dir = spod_get_data_dir(),
  quiet = FALSE,
  read_fun = duckdb::tbl_file
)

Arguments

zones

The zones for which to download the data. Can be "districts" (or "dist", "distr") or "municipalities" (or "muni", "municip") for v1 data. Additionaly, these can be "large_urban_areas" (or "lau") for v2 data.

dates

A character or Date vector of dates to process. Kindly keep in mind that v1 and v2 data follow different data collection methodologies and may not be directly comparable. Therefore, do not try to request data from both versions for the same date range. If you need to compare data from both versions, please refer to the respective codebooks and methodology documents. The v1 data covers the period from 2020-02-14 to 2021-05-09, and the v2 data covers the period from 2022-01-01 to the present until further notice. The true dates range is checked against the available data for each version on every function run.

The possible values can be any of the following:

A single date in ISO (YYYY-MM-DD) or YYYYMMDD format. character or Date object.

A vector of dates in ISO (YYYY-MM-DD) or YYYYMMDD format. character or Date object. Can be any non-consecutive sequence of dates.

A date range

eigher a character or Date object of length 2 with clearly named elements start and end in ISO (YYYY-MM-DD) or YYYYMMDD format. E.g. c(start = "2020-02-15", end = "2020-02-17");

or a character object of the form YYYY-MM-DD_YYYY-MM-DD or YYYYMMDD_YYYYMMDD. For example, 2020-02-15_2020-02-17 or 20200215_20200217.

A regular expression to match dates in the format YYYYMMDD. character object. For example, ^202002 will match all dates in February 2020.

data_dir

The directory where the data is stored. Defaults to the value returned by spod_get_data_dir() which returns the value of the environment variable SPANISH_OD_DATA_DIR or a temporary directory if the variable is not set.

quiet

Logical. If TRUE, the function does not print messages to the console. Defaults to FALSE.

read_fun

The function to read the data. Defaults to duckdb::tbl_file.

Value

A tibble with the origin-destination data.