Load the origin-destination v1 data (2020-2021) for specified dates

This function retrieves the v1 (2020-2021) origin_destination_data for the specified dates. It checks if the requested data is already cached locally and downloads it if it is not. When all the requested data is cached, it creates a DuckDB connection to the cache data folder and provides an table

Usage

spod_get_od_v1(
  zones = c("districts", "dist", "distr", "distritos", "municipalities", "muni",
    "municip", "municipios"),
  dates = NULL,
  data_dir = spod_get_data_dir(),
  quiet = FALSE,
  duck_max_mem = 2,
  duck_max_threads = parallelly::availableCores()
)

Arguments

zones

The zones for which to download the data. Can be "districts" (or "dist", "distr", or the original Spanish "distritos") or "municipalities" (or "muni", "municip", or the original Spanish "municipios"). Additionaly, these can be "large_urban_areas" (or "lau", or the original Spanish "grandes_areas_urbanas", or "gau") for v2 data.

dates

A character or Date vector of dates to process. Kindly keep in mind that v1 and v2 data follow different data collection methodologies and may not be directly comparable. Therefore, do not try to request data from both versions for the same date range. If you need to compare data from both versions, please refer to the respective codebooks and methodology documents. The v1 data covers the period from 2020-02-14 to 2021-05-09, and the v2 data covers the period from 2022-01-01 to the present until further notice. The true dates range is checked against the available data for each version on every function run.

The possible values can be any of the following:

A single date in ISO (YYYY-MM-DD) or YYYYMMDD format. character or Date object.

A vector of dates in ISO (YYYY-MM-DD) or YYYYMMDD format. character or Date object. Can be any non-consecutive sequence of dates.

A date range

eigher a character or Date object of length 2 with clearly named elements start and end in ISO (YYYY-MM-DD) or YYYYMMDD format. E.g. c(start = "2020-02-15", end = "2020-02-17");

or a character object of the form YYYY-MM-DD_YYYY-MM-DD or YYYYMMDD_YYYYMMDD. For example, 2020-02-15_2020-02-17 or 20200215_20200217.

A regular expression to match dates in the format YYYYMMDD. character object. For example, ^202002 will match all dates in February 2020.

data_dir

The directory where the data is stored. Defaults to the value returned by spod_get_data_dir() which returns the value of the environment variable SPANISH_OD_DATA_DIR or a temporary directory if the variable is not set.

quiet

Logical. If TRUE, the function does not print messages to the console. Defaults to FALSE.

duck_max_mem

The maximum memory to use in GB. A conservative default is 3 GB, which should be enough for resaving the data to DuckDB form a folder of CSV.gz files while being small enough to fit in memory of most even old computers. For data analysis using the already converted data (in DuckDB or Parquet format) or with the raw CSV.gz data, it is recommended to increase it according to available resources.

duck_max_threads

The maximum number of threads to use. Defaults to the number of available cores minus 1.

Value

A DuckDB table connection object. It can be manupulated using dplyr verbs, or can be loaded into memory using dplyr::collect(). The structure of the object is as follows:

full_date

Date. The full date of the trip, including year, month, and day.

id_origin

factor. The identifier for the origin location of the trip, formatted as a code (e.g., '01001_AM').

id_destination

factor. The identifier for the destination location of the trip, formatted as a code (e.g., '01001_AM').

activity_origin

factor. The type of activity at the origin location (e.g., 'home', 'work').

activity_destination

factor. The type of activity at the destination location (e.g., 'home', 'other').

residence_province

factor. The province of residence for the individual making the trip (e.g. 'Cuenca', 'Girona').

time_slot

integer. The time slot during which the trip started, represented as an integer (e.g., 0, 1, 2).

distance

factor. The distance category of the trip, represented as a code (e.g., '002-005' for 2-5 km).

n_trips

double. The number of trips taken within the specified time slot and distance.

trips_total_length_km

double. The total length of all trips in kilometers for the specified time slot and distance.

year

double. The year of the trip.

month

double. The month of the trip.

day

double. The day of the trip.

This object also contains the reference to the source DuckDB conneciton with the full view of the cached data. It can be accessed using od_table$src$con. See examples below. The connection includes two views:

od_csv_raw - a raw table view of all cached CSV files with the origin-destination data that has been previously cached in $SPANISH_OD_DATA_DIR

od_csv_clean - a cleaned-up table view of od_csv_raw with column names and values translated and mapped to English. This still includes all cached data.

View od_csv_clean has the same structure as the filtered view 'od_filtered', which is returned by spod_get_od_v1() as a DuckDB table connection object. The view od_csv_raw has original Spanish column names and values and has the following structure:

fecha

Date. The date of the trip, including year, month, and day.

origen

character. The identifier for the origin location of the trip, formatted as a character string (e.g., '01001_AM').

destino

character. The identifier for the destination location of the trip, formatted as a character string (e.g., '01001_AM').

actividad_origen

character. The type of activity at the origin location (e.g., 'casa', 'trabajo').

actividad_destino

character. The type of activity at the destination location (e.g., 'otros', 'trabajo').

residencia

character. The code representing the residence of the individual making the trip (e.g., '01') according to the official INE classification.

edad

character. The age of the individual making the trip. This data is actaully filled with 'NA' values, which is why this column is removed in the cleaned-up and translated view described above.

periodo

integer. The time period during which the trip started, represented as an integer (e.g., 0, 1, 2).

distancia

character. The distance category of the trip, represented as a character string (e.g., '002-005' for 2-5 km).

viajes

double. The number of trips taken within the specified time period and distance.

viajes_km

double. The total length of all trips in kilometers for the specified time period and distance.

day

double. The day of the trip.

month

double. The month of the trip.

year

double. The year of the trip.

Examples

if (FALSE) { # \dontrun{

# create a connection to the v1 data
Sys.setenv(SPANISH_OD_DATA_DIR = "~/home/nosync/cache/mitma")
dates <- c("2020-02-14", "2020-03-14", "2021-02-14", "2021-02-14", "2021-02-15")
od_dist <- spod_get_od_v1(zones = "distr", dates = dates)

# od dist is a table view filtered to the specified dates

# access the source connection with all dates
# list tables
DBI::dbListTables(od_dist$src$con)
} # }